Refactor initializers in heat model
The *heat_from_defaults* and *heat_from_input_file* functions are
now void functions. They take a pointer to a pointer to a HeatModel.
If an allocated HeatModel is passed in, it's initialized. If a null
pointer is passed in, a HeatModel is allocated and is set as the
pointer target.